Revert SharpGLTF cabinet loader
