+++ WindowsNetProjects/OasisEditor/OasisEditor/OasisEditor.csproj
@@ -13,7 +13,6 @@
     <PackageReference Include="Dirkster.AvalonDock" Version="4.72.1" />
     <PackageReference Include="Dirkster.AvalonDock.Themes.VS2013" Version="4.72.0" />
     <PackageReference Include="HelixToolkit.Wpf" Version="2.27.1" />
-    <PackageReference Include="SharpGLTF.Toolkit" Version="1.0.6" />
     <PackageReference Include="PixiEditor.ColorPicker" Version="3.4.2.3" />
     <PackageReference Include="SkiaSharp" Version="2.88.8" />
     <PackageReference Include="SkiaSharp.Views.WPF" Version="2.88.8" />
